requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.1
Django==4.1.7
django-constance==2.9.1
django-cors-headers==3.14.0
django-filter==22.1
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
kombu==5.2.4
phonenumbers==8.13.6
prompt-toolkit==3.0.37
psycopg2-binary
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7.1
simplejwt==2.0.1
six==1.16.0
sqlparse==0.4.3
typing==3.7.4.3
vine==5.0.0
wcwidth==0.2.6
django-picklefield==3.1
gunicorn==20.1.0
pyotp